Remove the breadcrumb title spacing functionality
This removes the space between each breadcrumb title part when passing multiple positional arguments.

This is a better default since spaces can be added very easily but to remove them the user has to use the `{{concat}}` helper. This also makes it more consistent with how the `{{concat}}` and `{{page-title}}` helpers works.